requirements.txt
bottle==0.12.8
BottleDaemon==0.1.1
Pillow==2.9.0
python-dateutil==1.5
pyyaml==3.10
pandas==0.21.0
xlrd==1.1.0
